Updated models, services and optimise queries

+++ app/services/order/place.js
@@ -4,7 +4,10 @@ const db = require('@server/sequelizeNew')
 
 const constraints = {
   cartId: {
-    presence: { allowEmpty: false }
+    presence: { allowEmpty: false },
+    numericality: {
+      onlyInteger: true
+    }
   }
 }
 
@@ -14,7 +17,6 @@ class PlaceOrder extends ServiceBase {
   }
 
   async run () {
-    const Op            = Sequelize.Op;
     
     const cart = await db.cart.findOne({
       where: {
@@ -22,51 +24,67 @@ class PlaceOrder extends ServiceBase {
       }
     });
 
+    if(!cart) {
+      this.addError('cart','Cart id does not exists.');
+      return;
+    }
+
     const items = await db.cartItem.findAll({
       where: {
         cart_id: this.cartId
       }
     });
     
-    const order = await db.order.create({
-      user_id: cart.user_id,
-      coupon_id: cart.coupon_id,
-      shipping_address_id: cart.shipping_address_id,
-      billing_address_id: cart.billing_address_id,
-      discount: cart.discount,
-      taxes: cart.taxes,
-      shipping_amount: cart.shipping_amount,
-      total_amount: cart.total_amount,
-      status: 'Pending',
-      contact_no: cart.contact_no
-    })
+    const t = await db.sequelize.transaction();
+    try {
+      const order = await db.order.create({
+        user_id: cart.user_id,
+        coupon_id: cart.coupon_id,
+        shipping_address_id: cart.shipping_address_id,
+        billing_address_id: cart.billing_address_id,
+        discount: cart.discount,
+        taxes: cart.taxes,
+        shipping_amount: cart.shipping_amount,
+        total_amount: cart.total_amount,
+        status: 'Pending',
+        contact_no: cart.contact_no
+      }, {transaction : t})
 
-    const order_id = order.id;
+      const order_id = order.id;
 
-    if(items.length > 0) {
-      for(var i=0; i<=items.length-1; i++) {
-        await db.orderItem.create({
-          product_id: items[i]['product_id'],
-          order_id: order_id,
-          quantity: items[i]['quantity'],
-          price: items[i]['price']
-        })
+      let orderItems = [];
+      if(items.length > 0) {
+        for(var i=0; i<=items.length-1; i++) {
+          orderItems.push({
+            product_id: items[i]['product_id'],
+            order_id: order_id,
+            quantity: items[i]['quantity'],
+            price: items[i]['price']
+          })
+        }
       }
+
+      await db.orderItem.bulkCreate(orderItems,{transaction : t})
+
+      await db.cartItem.destroy({
+        where : {
+          cart_id: this.cartId
+        }
+      }, {transaction : t});
+
+      await db.cart.destroy({
+        where : {
+          id: this.cartId
+        }
+      }, {transaction : t});
+      
+      t.commit();
+      return { message : "Order placed successfully.", orderId: order_id }
+    } catch (err) {
+      t.rollback();
+      this.addError('cart',err);
+      return;
     }
-
-    await db.cartItem.destroy({
-      where : {
-        cart_id: this.cartId
-      }
-    });
-
-    await db.cart.destroy({
-      where : {
-        id: this.cartId
-      }
-    });
-
-    return { message : "Order placed successfully." }
   }
 }
